refactor: 인수테스트 베이스 클래스 리팩토링
